requirements.txt
# python==3.11.7
pip==24.0
setuptools==65.5.0
Django==5.0.4
djangorestframework==3.15.1
firebase-admin==6.5.0
django-storages==1.14.3
boto3==1.34.102
pillow==10.3.0
python-dotenv==1.0.1
dj-database-url==2.1.0
psycopg2-binary==2.9.9
Faker==25.2.0
pandas==2.2.2
scikit-learn==1.4.2
scikit-surprise==1.1.4
django-cors-headers==4.3.1
gunicorn==22.0.0
whitenoise==6.6.0
